safari-600.7-branch/LayoutTests/media/audio-play-event-expected.txt

Test that a 'play' event listener is triggered when fired by a new audio element.

RUN(mediaElement.src = findMediaFile('audio', 'content/test'))
RUN(mediaElement.volume = 1)
RUN(mediaElement.play())
EVENT(play)
EVENT(loadedmetadata)
EVENT(canplay)
EVENT(playing)
END OF TEST
